requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==3.0.10
qrcode==7.3.1
geopy==2.2.0
python-dotenv==0.19.0
supabase==1.0.3
geocoder==1.38.1 